# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "CRM",
    "description": "CRM API Documentation",
    "version": "1.0.0"
  },
  "host": "localhost:2023",
  "basePath": "/api/user/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/all": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "serviceProvider",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/updateStatus": {
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "example": "any"
                },
                "updatedBy": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceProvider": {
                  "example": "any"
                },
                "id": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/userDetails": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/verifyOtp": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "otp": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/resetPassword": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "reEnterPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/batchNvEligibiltyCheck": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/bulkActivatedUpload": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "uploadedBy": {
                  "example": "any"
                },
                "serviceProvider": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/bulkPostpaidActivatedUpload": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "uploadedBy": {
                  "example": "any"
                },
                "serviceProvider": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/verifyZip": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceProvider": {
                  "example": "any"
                },
                "carrier": {
                  "example": "any"
                },
                "zipCode": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/PWGverifyZip": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceProvider": {
                  "example": "any"
                },
                "carrier": {
                  "example": "any"
                },
                "zipCode": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "accountType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/DBverifyzip": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceProvider": {
                  "example": "any"
                },
                "carrier": {
                  "example": "any"
                },
                "zipCode": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "accountType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/withoutzip": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "serviceProvider": {
                  "example": "any"
                },
                "carrier": {
                  "example": "any"
                },
                "csr": {
                  "example": "any"
                },
                "zipCode": {
                  "example": "any"
                },
                "department": {
                  "example": "any"
                },
                "accountType": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/initialInformation": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "firstName": {
                  "example": "any"
                },
                "middleName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "suffix": {
                  "example": "any"
                },
                "SSN": {
                  "example": "any"
                },
                "DOB": {
                  "example": "any"
                },
                "contact": {
                  "example": "any"
                },
                "isDifferentPerson": {
                  "example": "any"
                },
                "BenifitFirstName": {
                  "example": "any"
                },
                "BenifitMiddleName": {
                  "example": "any"
                },
                "BenifitLastName": {
                  "example": "any"
                },
                "BenifitSSN": {
                  "example": "any"
                },
                "BenifitDOB": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/homeAddress": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "city": {
                  "example": "any"
                },
                "address1": {
                  "example": "any"
                },
                "address2": {
                  "example": "any"
                },
                "state": {
                  "example": "any"
                },
                "zip": {
                  "example": "any"
                },
                "isTerribleTerritory": {
                  "example": "any"
                },
                "isBillAddress": {
                  "example": "any"
                },
                "mallingAddress1": {
                  "example": "any"
                },
                "mallingAddress2": {
                  "example": "any"
                },
                "mallingZip": {
                  "example": "any"
                },
                "mallingState": {
                  "example": "any"
                },
                "mallingCity": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/selectProgram": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "program": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/question": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "csr": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "livesWithAnotherAdult": {
                  "example": "any"
                },
                "hasAffordableConnectivity": {
                  "example": "any"
                },
                "isSharesIncomeAndExpense": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/termsAndConditions": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/plan": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "example": "any"
                },
                "plan": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/checkCustomerDuplication": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "alternateContact": {
                  "example": "any"
                },
                "contact": {
                  "example": "any"
                },
                "accountType": {
                  "example": "any"
                },
                "customerId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/postpaidpaymentDetails": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "customerId": {
                  "example": "any"
                },
                "planId": {
                  "example": "any"
                },
                "invoiceType": {
                  "example": "any"
                },
                "productName": {
                  "example": "any"
                },
                "selectProduct": {
                  "example": "any"
                },
                "planCharges": {
                  "example": "any"
                },
                "additionalCharges": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "totalAmount": {
                  "example": "any"
                },
                "amountPaid": {
                  "example": "any"
                },
                "invoiceDueDate": {
                  "example": "any"
                },
                "billingPeriod": {
                  "example": "any"
                },
                "invoiceStatus": {
                  "example": "any"
                },
                "paymentMethod": {
                  "example": "any"
                },
                "invoiceOneTimeCharges": {
                  "example": "any"
                },
                "lateFee": {
                  "example": "any"
                },
                "chargingType": {
                  "example": "any"
                },
                "printSetting": {
                  "example": "any"
                },
                "planName": {
                  "example": "any"
                },
                "billingModel": {
                  "example": "any"
                },
                "paymentChannel": {
                  "example": "any"
                },
                "paymentId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/handOverEnrollment": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "csr": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/prepaidHandOver": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "csr": {
                  "example": "any"
                },
                "userId": {
                  "example": "any"
                },
                "isWithoutInvoice": {
                  "example": "any"
                },
                "isWithInvoice": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/completeEnrollmentUser": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/prePostCompletedEnrollmentsList": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "accountType",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/prePostEvaluatedEnrollmentsList": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "accountType",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
    "/provisionedEnrollmentUserList": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "accountType",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
...
  }
}